move auth service in its own folder
