+++ frontend/src/app/core/auth/services/auth/auth.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
-import { environment } from '../../../../../environments/environment.development';
+import { environment } from '../../../../../../environments/environment.development';
 
 const api = environment.apiUrl;
 @Injectable({
